+++ packages/spruce-next-helpers/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.17.10](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.17.9...v8.17.10) (2019-11-15)
+
+
+### Bug Fixes
+
+* **npm-Cookies:** Update to version supporting sameSite [SDEV3-2622] ([bf9def6](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/bf9def6))
+
+
+
+
+
 ## [8.17.9](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.17.8...v8.17.9) (2019-11-05)
 
 **Note:** Version bump only for package @sprucelabs/spruce-next-helpers